% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonics_fun_eq.R
\name{harmonics_fun_eq}
\alias{harmonics_fun_eq}
\title{Equidistant time series modelling}
\usage{
harmonics_fun_eq(
  user_vals,
  user_dates,
  harmonic_deg = 3,
  ref_date = as.Date("1970-01-01", format = "\%Y-\%m-\%d"),
  new_dates = NULL,
  return_vals = "fitted",
  print_variance = FALSE,
  maxgap = NULL
)
}
\arguments{
\item{user_vals}{A vector with numeric values.}

\item{user_dates}{A Vector with Date objects. See \code{\link[base]{as.Date}}.}

\item{harmonic_deg}{Numeric. The number of cycles per year (harmonic degree)
that should be modelled.}

\item{ref_date}{(optional) A Date object. Default is 1970-01-01.}

\item{new_dates}{(optional) A vector with Date objects. See \code{\link[base]{as.Date}}. These are the desired artifical time steps.}

\item{return_vals}{(optional) String. Defines returned variables as one of "fitted","trend", "h1", "h2", (...and so on for chosen number of harmonics) or "all" to get a list of all of them. Default is "fitted".}

\item{print_variance}{(optional) Boolean. If TRUE, prints explained variance of every harmonic. Default is FALSE.}
}
\value{
A numeric vector with the fitted values or a list of vectors with fitted values, trend and all individual harmonics (in this order).
}
\description{
This function enables the user to model different number of cycles per year and calculate a desired number of artificial, equidistant time steps.
}
\details{
To calculate the harmonic fitted curve of a periodic signal,
         ordinary least squares regressions are computed using coupled
         sine and cosine curves on time-series data. The underlying algorithm
         is based on Shumway & Stoffer (2017) equations 4.1 – 4.2. Based on
         the derived function,the input of new dates allows to calculate
         artificial time steps other than the input dates.
}
\examples{

}
\references{
Shumway, R. H., & Stoffer, D. S. (2017). Time series analysis and its
            applications: with R examples. Springer.
}
