% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rHarmonics.R
\docType{package}
\name{rHarmonics}
\alias{rHarmonics}
\alias{rHarmonics-package}
\title{R package harmonic modelling}
\description{
This R package enables the user to model fitted harmonic values of various harmonic degrees.
}
\details{
Functions included in this package are:
\itemize{
  \item \code{\link[rHarmonics]{harmonics_fun}}
}
}
\author{
Marius Philipp \email{marius.philipp@uni-wuerzburg.de}
}
